fix(cli): stdout out of order for Confirm
